added possibility for electron dicom conversion
