+++ js/core/bis_commandlineutils.js
