+++ js/modules/bis_bidsmodule.js
