+++ js/modules/moduleindex.js
@@ -65,6 +65,11 @@ const moduleImports = {
     defaceimage:require('defaceImage.js'),
     reorientimage:require('reorientImage.js'),
     tfrecon:require('tfRecon.js'),
+
+    // these are here so that can be accessed by electron
+    'dicomconversion' : require('./dicommodule.js'),
+    'bidsconversion' : require('./bis_bidsmodule.js'),
+
 };
 
 // --------------------------------------------------------------
