+++ js/node/nodemoduleindex_base.js
@@ -22,9 +22,7 @@ const universalmodules=require('moduleindex');
 
 const moduleImports = {
     'infomodule': require('./infomodule.js'),
-    'dicomconversion' : require('./dicommodule.js'),
     'makechecksum' : require('./bis_hashmodule.js'),
-    'bidsconversion' : require('./bis_bidsmodule.js'),
     'regressiontests': require('./regressiontestmodule.js'),
 };
